% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic_tests.R
\name{diag.test.2mu}
\alias{diag.test.2mu}
\title{Diagnostic Test for Regime-Switching OUSSM}
\usage{
diag.test.2mu(ytn, tn.diff, split.pt, sim.num = 100)
}
\arguments{
\item{ytn}{Numeric vector. Observed time series.}

\item{tn.diff}{Numeric vector. Time differences.}

\item{split.pt}{Integer. Index of the regime shift.}

\item{sim.num}{Integer. Number of bootstrap simulations.}
}
\value{
A list containing test results and residual diagnostics.
}
\description{
Performs a bootstrap LRT to compare the Regime-Switching OUSSM against
a Regime-Switching Local Level model.
}
