% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic_tests.R
\name{diag.test}
\alias{diag.test}
\title{Diagnostic Test for OUSSM vs Local Level Model}
\usage{
diag.test(ytn, tn.diff, sim.num = 100)
}
\arguments{
\item{ytn}{Numeric vector. Observed time series.}

\item{tn.diff}{Numeric vector. Time differences.}

\item{sim.num}{Integer. Number of bootstrap simulations for the LRT.}
}
\value{
A list containing:
\item{test.result}{Boolean. TRUE if OUSSM is significantly better.}
\item{observed.logL.diff}{Numeric. The observed difference in log-likelihoods.}
\item{oussm.p}{Numeric. The p-value of the bootstrap test.}
\item{shapiro.p_value}{Numeric. P-value of the residual normality test.}
\item{vtn}{Vector. Standardized residuals.}
}
\description{
Performs a bootstrap Likelihood Ratio Test (LRT) to compare the OUSSM against
a Local Level (Random Walk + Noise) model. Also performs a Shapiro-Wilk test
on the standardized residuals.
}
