% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{estimation.OU.H.fixed.2mu}
\alias{estimation.OU.H.fixed.2mu}
\title{Estimate OUSSM Parameters with Regime-Switching Mean (Fixed H)}
\usage{
estimation.OU.H.fixed.2mu(theta.init, ytn, N, tn.diff, split.pt, H)
}
\arguments{
\item{theta.init}{Numeric. Initial guess for theta.}

\item{ytn}{Numeric vector. Observed time series.}

\item{N}{Integer. Number of observations.}

\item{tn.diff}{Numeric vector. Time differences.}

\item{split.pt}{Integer. Regime shift index.}

\item{H}{Numeric. The fixed observation variance.}
}
\value{
A list containing estimated parameters (mu1, mu2, theta, sigma) and optimization details.
}
\description{
Performs MLE for Regime-Switching OUSSM assuming the observation variance H is known/fixed.
}
