% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{estimation.OU.H.fixed}
\alias{estimation.OU.H.fixed}
\title{Estimate OUSSM Parameters (Fixed H)}
\usage{
estimation.OU.H.fixed(
  theta.init,
  ytn,
  N,
  tn.diff,
  H,
  Sigma.par.init = NULL,
  mu.init = NULL
)
}
\arguments{
\item{theta.init}{Numeric. Initial guess for theta.}

\item{ytn}{Numeric vector. Observed time series.}

\item{N}{Integer. Number of observations.}

\item{tn.diff}{Numeric vector. Time differences.}

\item{H}{Numeric. The fixed observation variance.}

\item{Sigma.par.init}{Numeric (optional). Initial guess for sigma.}

\item{mu.init}{Numeric (optional). Initial guess for mu.}
}
\value{
A list containing estimated parameters (mu, theta, sigma) and optimization details.
}
\description{
Performs MLE for OUSSM assuming the observation variance H is known/fixed.
}
