% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegLogL.R
\name{neglogL}
\alias{neglogL}
\title{Calculate Negative Log-Likelihood for OUSSM}
\usage{
neglogL(para, N, ytn, tn.diff)
}
\arguments{
\item{para}{A numeric vector of parameters to be estimated:
\itemize{
\item \code{para[1]}: log(theta)
\item \code{para[2]}: log(H)
\item \code{para[3]}: log(sigma)
\item \code{para[4]}: mu
}}

\item{N}{Integer. The total number of observations.}

\item{ytn}{Numeric vector. The observed values.}

\item{tn.diff}{Numeric vector. Time differences between observations.}
}
\value{
Numeric. The negative log-likelihood value.
}
\description{
This function calculates the negative log-likelihood of the Ornstein-Uhlenbeck
State-Space Model using the Kalman Filter.
}
